- Change extban_format() to size_t
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/branches/8.2.x@10041 82007160-df01-0410-b94d-b575c5fd34c7

+++ include/extban.h
@@ -70,7 +70,7 @@ extern void extban_del(struct Extban *);
 extern struct Extban *extban_find(unsigned char);
 extern struct Extban *extban_find_flag(unsigned int);
 extern enum extban_type extban_parse(const char *, unsigned int *, unsigned int *);
-extern unsigned int extban_format(unsigned int, char *);
+extern size_t extban_format(unsigned int, char *);
 extern unsigned int extban_matching_mask(void);
 extern unsigned int extban_acting_mask(void);
 extern const char *extban_get_isupport(void);
